Start coding the game 20220121 Now that i finally  got the set up of GPGS leaderboards over with i can start, really want to keep this simple so that i actually finish it. Tested it in desktop mode, it crashed because gsclient is null, that must be because gsclient is set in the android launcher, so this is what that ‘if gsclient== null’ code was for i guess, uncommented that, great, so it works in desktop mode,
For simplicity i’ll just put a stage over the GPGS menu and comment out where it is drawn in render, and don’t give it an inputprocessor
So i’ve got a blank screen, add an arraylist of symbols (derived from actors) for the symbols.
this arraylist can vary in size so it will grow as more symbols are added.
a symbol will be a class that inherits from actor. it will take an id, this id will determine how it looks. there will be a database of these ids to pull values from. it then needs an area id, either 0 or 1 that will tell us to display it in the top or bottom half of the screen, then a position id which will determine it's position in that area.
when displaying all the symbols i will display all the symbols in the bottom area, and all the symbols in the top area, then choose 1 symbol at random in the top area and change it's symbol id to the same value as a random symbol in the bottom area.
i've assigned the random symbolIds, seems to work, now i need to work on adding these actors to the stage and displaying them on screen