added the background system inspired by conway's game of life. background made up of a grid of cells, the bottom row of cells are randomly set to alive, if alive they turn red in colour. all cells count how many neighbours are alive and become alive themselves depending on customisable parameters regarding number of alive neighbours, we can also customise which cells are regarded as neighbours, so we can only consider cells below the current cell as neighoburs, this results in a sort of fire effect, the flame propagates upwards if there are enough red 'flame' cells alive below it.
i've already experimented and have picked out some combinations of neighbour, and how many neighbours need to be alive etc,that result in a good effect. some make it look like wind is blowing the fire from right to left or vice versa to different degrees, and the height and nature of the flame is increased with some combination, i want to utilise these different effects depending on the progress of the game.